kernelci.lab.LabAPI: pass job_path to .submit()
Rather than passing a job definition object to LabAPI.submit(), pass
the path to a job definition file.  Then it's up to the implementation
code to open it and submit it.  All the current use-cases are based on
actual definition files; if being able to pass an object instead
becomes required at some point then the class can be adjusted
accoringly then.

Update kernelci.lab.lava and kci_test accordingly.

Signed-off-by: Guillaume Tucker <guillaume.tucker@collabora.com>